projects/deploy_mnist/requirements.txt
Flask==1.0.2
numpy==1.16.2
#Pillow==2.2.1
scikit-learn==0.20.3
pillow==6.0.0
